♻️ help_surface 를 agents 로 이전해 discord 의존 제거
변경 이유:
agents → discord 순환 의존을 마무리하는 마지막 3개 import edge 중 하나로,
response_formatters 가 discord.engineering.help_surface 를 직접 import 해
순환을 만들고 있었다. help_surface 는 외부 의존이 0인 순수 텍스트 렌더러라
agents 쪽으로 옮기는 게 책임상 맞다.

주요 변경 사항:
- discord/engineering/help_surface.py 를
  agents/engineering_conversation/help_surface.py 로 git mv.
- response_formatters 가 로컬 .help_surface 에서 import 하도록 변경.
- 기존 discord 경로에는 sys.modules 별칭 shim 을 남겨
  discord/commands · bot/_legacy · 기존 테스트가 객체 동일성을 유지한 채
  forward 로 해소되게 함.

비고:
render_engineer_help_message / render_engineer_help_short public 이름은 그대로 보존.
